pyproject.toml
[project]
name = "robot-manipulator"
version = "0.1.0"
description = "SO101 arm lab — teleoperation, pick-and-place, imitation learning and beyond"
requires-python = ">=3.12"
dependencies = [
    # Robot arm control — Feetech servo support for SO101 leader/follower arms
    # OpenCV is bundled as a transitive dependency; do NOT add opencv-python here.
    "lerobot[feetech]>=0.5.0",
    # AI agent orchestration for zero-code task pipelines (https://cmdop.com)
    "openclaw>=2026.3.12",
    # Numerics / data
    "numpy>=2.0",
    "scipy>=1.14",
    # Config & environment
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "python-dotenv>=1.0",
    # Rich terminal output
    "rich>=13.9",
    # CLI
    "typer>=0.12",
]

[project.optional-dependencies]
dev = [
    "jupyterlab>=4.3",
    "ipywidgets>=8.1",
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
calibrate = "robot_manipulator.scripts.calibrate:main"
teleoperate = "robot_manipulator.scripts.teleoperate:main"
pick-and-place = "robot_manipulator.scripts.pick_and_place:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/robot_manipulator"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
